refactor(api): enhance logging system with better abstraction
- Created ILogger interface in application layer
- Implemented Logger in infrastructure layer
- Kept same logging functionality but with cleaner interfaces
- Updated all references to logger across the application
