+++ apps/express-app/src/application/providers/logger.interface.ts
@@ -1,0 +1,11 @@
+export interface LogMethod {
+  (message: string, meta?: any): void;
+}
+
+export interface ILogger {
+  error: LogMethod;
+  warn: LogMethod;
+  info: LogMethod;
+  http: LogMethod;
+  debug: LogMethod;
+}
