--- apps/express-app/src/services/logger/index.ts
@@ -1,3 +1,0 @@
-export * from "./logger.constants";
-export * from "./logger.service";
-export * from "./logger.types";
